Moving functions to utils file
